1.調整資料夾結構
