+++ app/build.gradle
@@ -43,4 +43,6 @@ dependencies {
 
     implementation "androidx.cardview:cardview:1.0.0"
     implementation 'com.github.bumptech.glide:glide:4.9.0'
+
+    implementation 'android.arch.paging:runtime:1.0.0'
 }
