+++ app/src/main/java/com/example/githubapisample/ui/collaborator/CollaboratorsFragment.java
@@ -1,9 +1,7 @@
-package com.example.githubapisample.ui;
+package com.example.githubapisample.ui.collaborator;
 
-import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
