+++ app/src/main/java/com/example/githubapisample/ui/collaborator/CollaboratorsViewModel.java
@@ -1,4 +1,4 @@
-package com.example.githubapisample.ui;
+package com.example.githubapisample.ui.collaborator;
 
 import android.text.TextUtils;
 
@@ -10,37 +10,37 @@ import androidx.lifecycle.ViewModel;
 
 import com.example.githubapisample.api.ApiResponse;
 import com.example.githubapisample.data.DataModel;
-import com.example.githubapisample.data.model.Commit;
+import com.example.githubapisample.data.model.Contributors;
 import com.example.githubapisample.util.AbsentLiveData;
 
 import java.util.List;
 
-public class CommitViewModel extends ViewModel {
+public class CollaboratorsViewModel extends ViewModel {
     private DataModel dataModel;
-    private final LiveData<ApiResponse<List<Commit>>> commit;
+    private final LiveData<ApiResponse<List<Contributors>>> collaborators;
     private final MutableLiveData<String> query = new MutableLiveData<>();
     private String userName;
 
-    public CommitViewModel(final DataModel dataModel) {
+    public CollaboratorsViewModel(final DataModel dataModel) {
         super();
         this.dataModel = dataModel;
-        commit = Transformations.switchMap(query, new Function<String, LiveData<ApiResponse<List<Commit>>>>() {
+        collaborators = Transformations.switchMap(query, new Function<String, LiveData<ApiResponse<List<Contributors>>>>() {
             @Override
-            public LiveData<ApiResponse<List<Commit>>> apply(String repo) {
+            public LiveData<ApiResponse<List<Contributors>>> apply(String repo) {
                 if (TextUtils.isEmpty(repo)) {
                     return AbsentLiveData.create();
                 } else {
-                    return dataModel.getCommitList(userName, repo);
+                    return dataModel.getContributorsList(userName, repo, 0);
                 }
             }
         });
     }
 
-    public LiveData<ApiResponse<List<Commit>>> getReposCommit() {
-        return commit;
+    public LiveData<ApiResponse<List<Contributors>>> getReposCollaborator() {
+        return collaborators;
     }
 
-    public void searchReposCommit(String userName, String repo) {
+    public void searchReposCollaborators(String userName, String repo) {
         this.userName = userName;
         query.setValue(repo);
     }
