+++ app/src/main/java/com/example/githubapisample/ui/commit/CommitFragment.java
@@ -1,8 +1,7 @@
-package com.example.githubapisample.ui;
+package com.example.githubapisample.ui.commit;
 
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
