+++ app/src/main/java/com/example/githubapisample/ui/repositories/UserRepoFragment.java
@@ -1,9 +1,8 @@
-package com.example.githubapisample.ui;
+package com.example.githubapisample.ui.repositories;
 
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -16,7 +15,6 @@ import androidx.lifecycle.ViewModelProviders;
 import androidx.recyclerview.widget.LinearLayoutManager;
 
 import com.example.githubapisample.DetailActivity;
-import com.example.githubapisample.R;
 import com.example.githubapisample.api.ApiResponse;
 import com.example.githubapisample.data.model.Repository;
 import com.example.githubapisample.databinding.UserRepoFragmentBinding;
