+++ app/src/main/res/layout/user_repo_collaborator.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="viewModel"
-            type="com.example.githubapisample.ui.CollaboratorsViewModel" />
+            type="com.example.githubapisample.ui.collaborator.CollaboratorsViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
